+++ src/app/app.module.ts
@@ -18,8 +18,12 @@ import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { HttpClientModule } from '@angular/common/http';
 import { CommonModule } from '@angular/common';
 import { RouterModule, Routes } from '@angular/router';
-import { BotonNotificacionComponent } from "./components/modules/boton-notificacion/boton-notificacion.component";
+
 import { UsuariosModule } from './components/modules/usuarios/usuarios.module';
+import { ListboxModule } from 'primeng/listbox';
+import { ButtonModule } from 'primeng/button';
+import { DialogModule } from 'primeng/dialog';
+import { DialogNnaMsgSeguimientoComponent } from './components/modules/dialog-nna-msg-seguimiento/dialog-nna-msg-seguimiento.component';
 
 @NgModule({
   declarations: [
@@ -29,6 +33,7 @@ import { UsuariosModule } from './components/modules/usuarios/usuarios.module';
     HeaderComponent,
     FooterComponent,
     ContentComponent,
+    
   ],
   imports: [
     UsuariosModule,
@@ -40,13 +45,17 @@ import { UsuariosModule } from './components/modules/usuarios/usuarios.module';
     CommonModule,
     RouterModule,
     FormsModule,
+
     /** PrimeNG */
     MenuModule,
     BadgeModule,
     RippleModule,
     AvatarModule,
     TableModule,
-    BotonNotificacionComponent
+
+    /**Component standalone */
+    
+    
 ],
   providers: [],
   bootstrap: [AppComponent]
