+++ src/app/components/modules/dialog-nna-msg-seguimiento/dialog-nna-msg-seguimiento.component.html
@@ -1,0 +1,75 @@
+<p-dialog header="" [(visible)]="visible" [modal]="true" [breakpoints]="{ '1199px': '75vw', '575px': '90vw' }"
+    [style]="{ width: '50vw' }" [draggable]="false" [resizable]="false">
+    <div class="col-sm-12 text-center">
+        <label>Agendar seguimiento</label>
+    </div>
+    <div class="col-sm-12 text-left">
+        <div class="row">
+            <div class="col-sm-3">
+                <label>Agente de seguimiento*:</label>
+            </div>
+            <div class="col-sm-9 ">
+                <select class="form-select" [ngModel]="formAgenteSeguimiento" id="Agentes" name="Agentes">
+                    <option selected value="">Seleccione</option>
+                    <option *ngFor="let agente of agenteAsignadoListado" [value]="agente.id">
+                        {{ agente.fullName }}
+                    </option>
+
+                </select>
+            </div>
+        </div>
+    </div>
+    <div class="col-sm-12 text-left">
+        <div class="row">
+            <div class="col-sm-3">
+                <label>Fecha*:</label>
+            </div>
+            <div class="col-sm-3">
+                <div class="row">
+                    <div class="col-sm-12">
+                        <input type="date" class="form-control" [(ngModel)]="formFecha" />
+                    </div>
+                    <div class="col-sm-12">
+                        <div class="row">
+                            <div class="col-sm-5">
+                                <button [class]="'btn text-color-blanco '+buttonHoy" (click)="cargarHoy()">Hoy</button>
+                            </div>
+                            <div class="col-sm-5">
+                                <button [class]="'btn text-color-blanco '+buttonManana"
+                                    (click)="cargarManana()">Mañana</button>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+            <div class="col-sm-1 text-right">
+                <label>Hora*:</label>
+            </div>
+            <div class="col-sm-5">
+                <div class="row">
+                    <div class="col-sm-3">
+                        <input type="number" class="form-control col-sm-2" placeholder="HH" [(ngModel)]="formHora" />
+                    </div>
+                    <div class="col-sm-3">
+                        <input type="number" class="form-control col-sm-2" placeholder="MM" [(ngModel)]="formMinuto" />
+                    </div>
+                    <div class="col-sm-2">
+                        <button [class]="'btn text-color-blanco '+buttonAm" (click)="cargarAmPM('AM')">AM</button>
+                    </div>
+                    <div class="col-sm-2">
+                        <button [class]="'btn text-color-blanco '+buttonPm" (click)="cargarAmPM('PM')">PM</button>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+    <div class="col-sm-12" *ngIf="!axios.isEmpty(msg)">
+        <br />
+        <small class="errors" *ngIf="!axios.isEmpty(msg)">{{msg}}</small>
+    </div>
+    <div class="col-sm-12 text-center">
+        <br />
+        <button class="btn text-color-blanco fondo-color-principal" (click)="guardar()">GUARDAR</button>
+    </div>
+
+</p-dialog>
