--- src/app/components/modules/usuarios/dialog-crear-nna-msg-rol-coordinador/dialog-crear-nna-msg-rol-coordinador.component.html
@@ -1,20 +1,0 @@
-<app-dialog-nna-msg-seguimiento [visible]="visibleDialogSeguimiento" [nnaId]="nnaId" [agenteId]="agenteId" [coordinadorId]="coordinadorId"></app-dialog-nna-msg-seguimiento>
-
-<p-dialog header="" [(visible)]="visible" [modal]="true" [breakpoints]="{ '1199px': '75vw', '575px': '90vw' }"
-    [style]="{ width: '50vw' }" [draggable]="false" [resizable]="false">
-    <div class="text-center">
-        <div class="col-sm-12">
-            <button class="btn text-color-blanco fondo-color-principal rondear-full">
-                <i class="pi pi-check"></i>
-            </button>
-        </div>
-        <div class="col-sm-12">
-            <label>NNA creado con éxito</label>
-        </div>
-        <div class="col-sm-6">
-            <button class="btn text-color-blanco fondo-color-principal" (click)="onSeguimiento()">
-                <span>ASIGNAR Y AGENDAR SEGUIMIENTO</span>
-            </button>
-        </div>        
-    </div>    
-</p-dialog>
