+++ src/app/core/services/generico.ts
@@ -113,4 +113,9 @@ export class Generico {
     const request = axios.post(url, data, { headers });
     return this.handleRequest(request);
   }
+
+  // Método para verificar si un campo está vacío
+  isEmpty(value: any): boolean {
+    return value === null || value === undefined || value.trim() === '';
+  }
 }
